+++ tests/hypothesis_strategies/kernels.py
@@ -16,7 +16,7 @@ from baybe.kernels.basic import (
 )
 from baybe.kernels.composite import AdditiveKernel, ProductKernel, ScaleKernel
 
-from ..hypothesis_strategies.basic import finite_floats
+from ..hypothesis_strategies.basic import positive_finite_floats
 from ..hypothesis_strategies.priors import priors
 
 
@@ -31,7 +31,7 @@ class KernelType(Enum):
 linear_kernels = st.builds(
     LinearKernel,
     variance_prior=st.one_of(st.none(), priors),
-    variance_initial_value=st.one_of(st.none(), finite_floats()),
+    variance_initial_value=st.one_of(st.none(), positive_finite_floats()),
 )
 """A strategy that generates linear kernels."""
 
@@ -39,16 +39,16 @@ matern_kernels = st.builds(
     MaternKernel,
     nu=st.sampled_from((0.5, 1.5, 2.5)),
     lengthscale_prior=st.one_of(st.none(), priors),
-    lengthscale_initial_value=st.one_of(st.none(), finite_floats()),
+    lengthscale_initial_value=st.one_of(st.none(), positive_finite_floats()),
 )
 """A strategy that generates Matern kernels."""
 
 periodic_kernels = st.builds(
     PeriodicKernel,
     lengthscale_prior=st.one_of(st.none(), priors),
-    lengthscale_initial_value=st.one_of(st.none(), finite_floats()),
+    lengthscale_initial_value=st.one_of(st.none(), positive_finite_floats()),
     period_length_prior=st.one_of(st.none(), priors),
-    period_length_initial_value=st.one_of(st.none(), finite_floats()),
+    period_length_initial_value=st.one_of(st.none(), positive_finite_floats()),
 )
 """A strategy that generates periodic kernels."""
 
@@ -56,7 +56,7 @@ piecewise_polynomial_kernels = st.builds(
     PiecewisePolynomialKernel,
     q=st.integers(min_value=0, max_value=3),
     lengthscale_prior=st.one_of(st.none(), priors),
-    lengthscale_initial_value=st.one_of(st.none(), finite_floats()),
+    lengthscale_initial_value=st.one_of(st.none(), positive_finite_floats()),
 )
 """A strategy that generates piecewise polynomial kernels."""
 
@@ -64,14 +64,14 @@ polynomial_kernels = st.builds(
     PolynomialKernel,
     power=st.integers(min_value=0),
     offset_prior=st.one_of(st.none(), priors),
-    offset_initial_value=st.one_of(st.none(), finite_floats()),
+    offset_initial_value=st.one_of(st.none(), positive_finite_floats()),
 )
 """A strategy that generates polynomial kernels."""
 
 rbf_kernels = st.builds(
     RBFKernel,
     lengthscale_prior=st.one_of(st.none(), priors),
-    lengthscale_initial_value=st.one_of(st.none(), finite_floats()),
+    lengthscale_initial_value=st.one_of(st.none(), positive_finite_floats()),
 )
 """A strategy that generates radial basis function (RBF) kernels."""
 
@@ -79,14 +79,14 @@ rff_kernels = st.builds(
     RFFKernel,
     num_samples=st.integers(min_value=1),
     lengthscale_prior=st.one_of(st.none(), priors),
-    lengthscale_initial_value=st.one_of(st.none(), finite_floats()),
+    lengthscale_initial_value=st.one_of(st.none(), positive_finite_floats()),
 )
 """A strategy that generates random Fourier features (RFF) kernels."""
 
 rq_kernels = st.builds(
     RQKernel,
     lengthscale_prior=st.one_of(st.none(), priors),
-    lengthscale_initial_value=st.one_of(st.none(), finite_floats()),
+    lengthscale_initial_value=st.one_of(st.none(), positive_finite_floats()),
 )
 """A strategy that generates rational quadratic (RQ) kernels."""
 
@@ -115,7 +115,7 @@ def single_kernels(draw: st.DrawFn):
             base_kernel=base_kernel,
             outputscale_prior=draw(st.one_of(st.none(), priors)),
             outputscale_initial_value=draw(
-                st.one_of(st.none(), finite_floats()),
+                st.one_of(st.none(), positive_finite_floats()),
             ),
         )
     else:
@@ -124,13 +124,13 @@ def single_kernels(draw: st.DrawFn):
 
 @st.composite
 def kernels(draw: st.DrawFn):
-    """Generate :class:`baybe.kernels.basic.Kernel`."""
+    """Generate :class:`baybe.kernels.base.Kernel`."""
     kernel_type = draw(st.sampled_from(KernelType))
 
     if kernel_type is KernelType.SINGLE:
         return draw(single_kernels())
 
-    base_kernels = draw(st.lists(single_kernels()))
+    base_kernels = draw(st.lists(single_kernels(), min_size=2))
     if kernel_type is KernelType.ADDITIVE:
         return AdditiveKernel(base_kernels)
     if kernel_type is KernelType.PRODUCT: